ADDED: completed all test cases with parameterized testing for feet and yard
