+++ quantity_custom_execption.py
@@ -1,11 +1,11 @@
 '''
 @author: Shivam Mishra
 @date: 29-12-21 12:21 PM
-
+@desc: Custom Quantity Exception
 '''
 
 
-class QuatityCustomException(Exception):
+class QuantityCustomException(Exception):
 
     def __init__(self,message):
         self.message = message
