+++ tests/test_quantity_mesurement.py
@@ -5,8 +5,8 @@
 '''
 import pytest as pytest
 
-from quatity_custom_execption import QuatityCustomException
-from quatity_mesurement import Measurement
+from quantity_custom_execption import QuantityCustomException
+from quantity_mesurement import Measurement
 
 
 class TestMeasurement:
@@ -21,7 +21,7 @@ class TestMeasurement:
     def test_if_none_return_exception(self, len1, unit1, len2, unit2):
         obj1 = Measurement(len1, unit1)
         obj2 = Measurement(len2, unit2)
-        with pytest.raises(QuatityCustomException) as exeception:
+        with pytest.raises(QuantityCustomException) as exeception:
             obj1 == obj2
         assert exeception.value.message == "Value can't be none"
 
@@ -29,7 +29,7 @@ class TestMeasurement:
     def test_whether_reference_are_equal(self, len1, unit1, len2, unit2):
         obj1 = Measurement(len1, unit1)
         obj2 = Measurement(len2, unit2)
-        with pytest.raises(QuatityCustomException) as exeception:
+        with pytest.raises(QuantityCustomException) as exeception:
             obj1 == obj2
         assert exeception.value.message == "Reference are different"
 
@@ -40,21 +40,17 @@ class TestMeasurement:
         assert obj1 == obj2
 
     @pytest.mark.parametrize('len1,unit1,len2,unit2', [(2.0, "feet", 2.0, "feet"), (2.0, "inch", 2.0, "inch"),
-                                                       (1.0, "feet", 12.0, "inch"),(1.0, "inch", 12.0, "feet")])
+                                                       (1.0, "feet", 12.0, "inch"),(1.0, "inch", 12.0, "feet"),
+                                                       (3.0, "feet", 1.0, "yard"),(1.0, "yard", 3.0, "feet"),
+                                                       (1.0, "yard", 36.0, "inch"), (36.0, "inch", 1.0, "yard" )])
     def test_whether_value_is_equal(self, len1, unit1, len2, unit2):
         obj1 = Measurement(len1, unit1)
         obj2 = Measurement(len2, unit2)
         assert obj1 == obj2
 
-    @pytest.mark.parametrize('len1,unit1,len2,unit2', [(1.0, "feet", 1.0, "inch"), (1.0, "inch", 1.0, "feet")])
+    @pytest.mark.parametrize('len1,unit1,len2,unit2', [(1.0, "feet", 1.0, "inch"), (1.0, "inch", 1.0, "feet"),
+                                                       (1.0, "feet", 1.0, "yard"), (1.0, "inch", 1.0, "yard")])
     def test_whether_values_are_not_equal(self, len1, unit1, len2, unit2):
         obj1 = Measurement(len1, unit1)
         obj2 = Measurement(len2, unit2)
         assert obj1 != obj2
-
-
-
-
-
-
-
